apps/requirements.txt
# Databricks SDK and SQL Connector
databricks-sdk>=0.20.0
databricks-sql-connector>=3.0.0

# Streamlit
streamlit>=1.30.0

# TabPFN Client
tabpfn-client>=0.1.0

# Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# PyArrow for efficient data transfer
pyarrow>=14.0.0
